Task: 7 Advanced use of web storage
Setup your files with the following:

Create an array availableItems that will contain all the available items. Add the strings Shampoo, Soap, Sponge, and Water to the array
If session storage is not enabled on your browser, display an alert that will contain the message Sorry, your browser does not support Web storage. Try again with a better one
If session storage is available it should allow the user to see the application and call the function createStore and displayCart

Create a function getCartFromStorage:
It should parse a string into a JSON object, returning the content of the cart stored in Session storage
If there is no cart, it should return an empty object

Create a function addItemToCart:
It accepts item(string) as argument
It adds to the cart object the item
If the same item is added multiple times, the cart store the quantity
It stores the value of the cart object in a string for the key cart in the Session Storage
It calls displayCart

Create a function removeItemfromCart:
It accepts item(string) as argument
It remove the entire item from the cart
Store the value of the cart object in a string for the key cart in the Session Storage
It calls displayCart

Create a function clearCart:
It should clear the entire Session storage
it calls displayCart

Create a function createStore:
It should add a h2 tag with the text Available products:
It should add a list with every item available for purchase
When the user click on an item, it should add it to the cart

Create a function displayCart:
It should add inside a h2 tag with the text Your cart:
It should add an empty div tag
If the div tag already exist, it should remove any list child
It calls updateCart

Create a function updateCart:
It should add a list to the div tag created previously
If the cart is empty, it should add an item Your cart is empty
If the cart is not empty, it should add the list of items within the cart with the following format: ITEM_NAME x QUANTITY (remove)
When the user clicks on remove, it should call the function removeItemfromCart
At the top of the cart, add an item named Clear my cart. When the user clicks on it, it should call the function clearCart